DEVELOPMENT_2026_02_11_2012. implementing JNI value mangling.

+++ Assignment_1_-_the_dll/LIB_Math.cpp
@@ -23,14 +23,14 @@
 		stat_CLASS_get_Framework()->initialise(stat_CLASS_get_Framework());
 		return (void*)stat_CLASS_get_Framework();
 	}
-	// get.
+		// get.
 	float CLIBAdder::LIB_Math::get_Input_Subset_praise0_valueA(void* obj)
 	{
-		return CLIBAdder::Global::obj_getClassOf(obj)->get_CLASS_Adder()->dyn_REG_get_Input_Subset_praise0_valueA(CLIBAdder::Global::obj_getClassOf(obj));
+		return static_cast<jfloat>(CLIBAdder::Global::obj_getClassOf(obj)->get_CLASS_Adder()->dyn_REG_get_Input_Subset_praise0_valueA(CLIBAdder::Global::obj_getClassOf(obj)));
 	}
 	float CLIBAdder::LIB_Math::get_Input_Subset_praise0_valueB(void* obj)
 	{
-		return CLIBAdder::Global::obj_getClassOf(obj)->get_CLASS_Adder()->dyn_REG_get_Input_Subset_praise0_valueB(CLIBAdder::Global::obj_getClassOf(obj));
+		return static_cast<jfloat>(CLIBAdder::Global::obj_getClassOf(obj)->get_CLASS_Adder()->dyn_REG_get_Input_Subset_praise0_valueB(CLIBAdder::Global::obj_getClassOf(obj)));
 	}
 		// set.
 	void CLIBAdder::LIB_Math::set_Input_Subset_praise0_valueA(void* obj, float newValue)
